desktop/requirements-desktop.txt
pywebview
uvicorn[standard]
